Struct wrapped2d::dynamics::body::MetaBody

pub struct MetaBody<U: UserDataTypes> { /* fields omitted */ }

Methods

impl<U: UserDataTypes> MetaBody<U>

pub fn create_fixture(
&mut self,
shape: &Shape,
def: &mut FixtureDef
) -> FixtureHandle where
U::FixtureData: Default,

pub fn create_fixture_with(
&mut self,
shape: &Shape,
def: &mut FixtureDef,
data: U::FixtureData
) -> FixtureHandle

pub fn create_fast_fixture(
&mut self,
shape: &Shape,
density: f32
) -> FixtureHandle where
U::FixtureData: Default,

pub fn create_fast_fixture_with(
&mut self,
shape: &Shape,
density: f32,
data: U::FixtureData
) -> FixtureHandle

pub fn fixture(&self, handle: FixtureHandle) -> Ref<MetaFixture<U>>

pub fn fixture_mut(&self, handle: FixtureHandle) -> RefMut<MetaFixture<U>>

pub fn destroy_fixture(&mut self, handle: FixtureHandle)

pub fn fixtures(&self) -> HandleIter<Fixture, MetaFixture<U>>

Important traits for ContactIter<'a>
impl<'a> Iterator for ContactIter<'a> type Item = (BodyHandle, WrappedRef<'a, Contact>);

pub fn contacts(&self) -> ContactIter

This method is here because contacts are owned by the world and not by the body, and having a reference to a MetaBody requires having a reference to the world.

Important traits for ContactIterMut<'a>
impl<'a> Iterator for ContactIterMut<'a> type Item = (BodyHandle, WrappedRefMut<'a, Contact>);

pub unsafe fn contacts_mut(&mut self) -> ContactIterMut

This method is here because contacts are owned by the world and not by the body, however having a mutable reference to a MetaBody only requires having an immutable reference to the World, that's why this method is unsafe.