Struct wrapped2d::dynamics::body::MetaBody

pub struct MetaBody<U: UserDataTypes> { /* fields omitted */ }

Implementations

impl<U: UserDataTypes> MetaBody<U>

pub fn create_fixture(
&mut self,
shape: &dyn Shape,
def: &mut FixtureDef
) -> FixtureHandle where
U::FixtureData: Default,

pub fn create_fixture_with(
&mut self,
shape: &dyn Shape,
def: &mut FixtureDef,
data: U::FixtureData
) -> FixtureHandle

pub fn create_fast_fixture(
&mut self,
shape: &dyn Shape,
density: f32
) -> FixtureHandle where
U::FixtureData: Default,

pub fn create_fast_fixture_with(
&mut self,
shape: &dyn Shape,
density: f32,
data: U::FixtureData
) -> FixtureHandle

pub fn fixture(&self, handle: FixtureHandle) -> Ref<'_, MetaFixture<U>>

pub fn fixture_mut(&self, handle: FixtureHandle) -> RefMut<'_, MetaFixture<U>>

pub fn destroy_fixture(&mut self, handle: FixtureHandle)

pub fn fixtures(&self) -> HandleIter<'_, Fixture, MetaFixture<U>>

pub fn contacts(&self) -> ContactIter<'_>

This method is here because contacts are owned by the world and not by the body, and having a reference to a MetaBody requires having a reference to the world.

pub unsafe fn contacts_mut(&mut self) -> ContactIterMut<'_>

This method is here because contacts are owned by the world and not by the body, however having a mutable reference to a MetaBody only requires having an immutable reference to the World, that's why this method is unsafe.